+++ example/lib/main.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:animated_bottom_navigation_bar/animated_bottom_navigation_bar.dart';
 import 'package:auto_size_text/auto_size_text.dart';
 import 'package:circular_reveal_animation/circular_reveal_animation.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 
 void main() => runApp(MyApp());
@@ -34,14 +34,17 @@ class MyHomePage extends StatefulWidget {
   _MyHomePageState createState() => _MyHomePageState();
 }
 
-class _MyHomePageState extends State<MyHomePage>
-    with SingleTickerProviderStateMixin {
+class _MyHomePageState extends State<MyHomePage> with TickerProviderStateMixin {
   final autoSizeGroup = AutoSizeGroup();
   var _bottomNavIndex = 0; //default index of a first screen
 
-  late AnimationController _animationController;
-  late Animation<double> animation;
-  late CurvedAnimation curve;
+  late AnimationController _fabAnimationController;
+  late AnimationController _borderRadiusAnimationController;
+  late Animation<double> fabAnimation;
+  late Animation<double> borderRadiusAnimation;
+  late CurvedAnimation fabCurve;
+  late CurvedAnimation borderRadiusCurve;
+  late AnimationController _hideBottomBarAnimationController;
 
   final iconList = <IconData>[
     Icons.brightness_5,
@@ -59,27 +62,60 @@ class _MyHomePageState extends State<MyHomePage>
     );
     SystemChrome.setSystemUIOverlayStyle(systemTheme);
 
-    _animationController = AnimationController(
-      duration: Duration(seconds: 1),
+    _fabAnimationController = AnimationController(
+      duration: Duration(milliseconds: 500),
       vsync: this,
     );
-    curve = CurvedAnimation(
-      parent: _animationController,
-      curve: Interval(
-        0.5,
-        1.0,
-        curve: Curves.fastOutSlowIn,
-      ),
+    _borderRadiusAnimationController = AnimationController(
+      duration: Duration(milliseconds: 500),
+      vsync: this,
     );
-    animation = Tween<double>(
-      begin: 0,
-      end: 1,
-    ).animate(curve);
+    fabCurve = CurvedAnimation(
+      parent: _fabAnimationController,
+      curve: Interval(0.5, 1.0, curve: Curves.fastOutSlowIn),
+    );
+    borderRadiusCurve = CurvedAnimation(
+      parent: _borderRadiusAnimationController,
+      curve: Interval(0.5, 1.0, curve: Curves.fastOutSlowIn),
+    );
+
+    fabAnimation = Tween<double>(begin: 0, end: 1).animate(fabCurve);
+    borderRadiusAnimation = Tween<double>(begin: 0, end: 1).animate(
+      borderRadiusCurve,
+    );
+
+    _hideBottomBarAnimationController = AnimationController(
+      duration: Duration(milliseconds: 200),
+      vsync: this,
+    );
 
     Future.delayed(
       Duration(seconds: 1),
-      () => _animationController.forward(),
+      () => _fabAnimationController.forward(),
     );
+    Future.delayed(
+      Duration(seconds: 1),
+      () => _borderRadiusAnimationController.forward(),
+    );
+  }
+
+  bool onScrollNotification(ScrollNotification notification) {
+    if (notification is UserScrollNotification &&
+        notification.metrics.axis == Axis.vertical) {
+      switch (notification.direction) {
+        case ScrollDirection.forward:
+          _hideBottomBarAnimationController.reverse();
+          _fabAnimationController.forward(from: 0);
+          break;
+        case ScrollDirection.reverse:
+          _hideBottomBarAnimationController.forward();
+          _fabAnimationController.reverse(from: 1);
+          break;
+        case ScrollDirection.idle:
+          break;
+      }
+    }
+    return false;
   }
 
   @override
@@ -95,23 +131,22 @@ class _MyHomePageState extends State<MyHomePage>
           ),
           backgroundColor: HexColor('#373A36'),
         ),
-        body: NavigationScreen(
-          iconList[_bottomNavIndex],
+        body: NotificationListener<ScrollNotification>(
+          onNotification: onScrollNotification,
+          child: NavigationScreen(iconList[_bottomNavIndex]),
         ),
-        floatingActionButton: ScaleTransition(
-          scale: animation,
-          child: FloatingActionButton(
-            elevation: 8,
-            backgroundColor: HexColor('#FFA400'),
-            child: Icon(
-              Icons.brightness_3,
-              color: HexColor('#373A36'),
-            ),
-            onPressed: () {
-              _animationController.reset();
-              _animationController.forward();
-            },
+        floatingActionButton: FloatingActionButton(
+          backgroundColor: HexColor('#FFA400'),
+          child: Icon(
+            Icons.brightness_3,
+            color: HexColor('#373A36'),
           ),
+          onPressed: () {
+            _fabAnimationController.reset();
+            _borderRadiusAnimationController.reset();
+            _borderRadiusAnimationController.forward();
+            _fabAnimationController.forward();
+          },
         ),
         floatingActionButtonLocation: FloatingActionButtonLocation.centerDocked,
         bottomNavigationBar: AnimatedBottomNavigationBar.builder(
@@ -143,13 +178,20 @@ class _MyHomePageState extends State<MyHomePage>
           backgroundColor: HexColor('#373A36'),
           activeIndex: _bottomNavIndex,
           splashColor: HexColor('#FFA400'),
-          notchAndCornersAnimation: animation,
+          notchAndCornersAnimation: borderRadiusAnimation,
           splashSpeedInMilliseconds: 300,
           notchSmoothness: NotchSmoothness.defaultEdge,
           gapLocation: GapLocation.center,
           leftCornerRadius: 32,
           rightCornerRadius: 32,
           onTap: (index) => setState(() => _bottomNavIndex = index),
+          hideAnimationController: _hideBottomBarAnimationController,
+          shadow: BoxShadow(
+            offset: Offset(0, 1),
+            blurRadius: 12,
+            spreadRadius: 0.5,
+            color: Colors.grey,
+          ),
         ),
       ),
     );
@@ -213,20 +255,23 @@ class _NavigationScreenState extends State<NavigationScreen>
   @override
   Widget build(BuildContext context) {
     return Container(
-      width: double.infinity,
-      height: double.infinity,
       color: Colors.white,
-      child: Center(
-        child: CircularRevealAnimation(
-          animation: animation,
-          centerOffset: Offset(80, 80),
-          maxRadius: MediaQuery.of(context).size.longestSide * 1.1,
-          child: Icon(
-            widget.iconData,
-            color: HexColor('#FFA400'),
-            size: 160,
+      child: ListView(
+        children: [
+          SizedBox(height: 64),
+          Center(
+            child: CircularRevealAnimation(
+              animation: animation,
+              centerOffset: Offset(80, 80),
+              maxRadius: MediaQuery.of(context).size.longestSide * 1.1,
+              child: Icon(
+                widget.iconData,
+                color: HexColor('#FFA400'),
+                size: 160,
+              ),
+            ),
           ),
-        ),
+        ],
       ),
     );
   }