+++ lib/src/circular_notch_and_corner_clipper.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import 'exceptions.dart';
-
 class CircularNotchedAndCorneredRectangleClipper extends CustomClipper<Path> {
   final ValueListenable<ScaffoldGeometry> geometry;
   final NotchedShape shape;
@@ -16,12 +14,6 @@ class CircularNotchedAndCorneredRectangleClipper extends CustomClipper<Path> {
 
   @override
   Path getClip(Size size) {
-    if (geometry.value.floatingActionButtonArea != null &&
-        geometry.value.floatingActionButtonArea!.width !=
-            geometry.value.floatingActionButtonArea!.height)
-      throw IllegalFloatingActionButtonSizeException(
-          'Floating action button must be a circle');
-
     final Rect? button = geometry.value.floatingActionButtonArea?.translate(
       0.0,
       geometry.value.bottomNavigationBarTop! * -1.0,